Fixes #234.
