+++ lib/styleset-utils.js
@@ -9,6 +9,7 @@ var copyStyle = require('../models/style.js').copy;
 var copyStyleValues = require('../models/style.js').copyValues;
 var isStylePopulated = require('../models/style.js').isPopulated;
 var font_utils = require('./font-utils');
+var cssNames = require(path.join(__dirname, 'css-names.json'));
 
 exports.getNonEditableCss = function (fontsInUse, next) {
 	var nonEditableCssFilename = path.resolve(__dirname, '../public/create/stylesets', 'non-editable.css');
@@ -197,4 +198,30 @@ exports.createStyle = function (name, clazz, css, tag, stylesetId, isSystem, hid
 		isSystem: isSystem,
 		hidden: hidden
 	});
+}
+
+function getCssKey ( cssNames, name ) {
+	for (var key in cssNames) {
+		if (cssNames.hasOwnProperty(key)) {
+			if(cssNames[key].name == name) {
+				return key;
+			}
+		}
+	}
+}
+
+/**
+ * Sort the styles by "system style order", i.e. so they appear in the same order as in "CSS names".
+ *
+ * @param s1
+ * @param s2
+ * @returns {number}
+ */
+exports.systemStyleOrder = function (s1, s2) {
+	var s1KeyName = getCssKey(cssNames, s1.name);
+	var s2KeyName = getCssKey(cssNames, s2.name);
+
+	var s1Order = cssNames[s1KeyName] ? cssNames[s1KeyName].order : 9999;
+	var s2Order = cssNames[s2KeyName] ? cssNames[s2KeyName].order : 9999;
+	return s1Order - s2Order;
 }
