+++ scripts/import_system_stylesets.js
@@ -10,32 +10,7 @@ var Style = require('../models/style.js').Style;
 var cssNames = require(path.join(__dirname, '../lib/css-names.json'));
 var filewalker = require('filewalker');
 var utils = require('../lib/utils');
-
-function getCssKey ( cssNames, name ) {
-	for (var key in cssNames) {
-		if (cssNames.hasOwnProperty(key)) {
-			if(cssNames[key].name == name) {
-				return key;
-			}
-		}
-	}
-}
-
-/**
- * Sort the styles by "system style order", i.e. so they appear in the same order as in "CSS names".
- *
- * @param s1
- * @param s2
- * @returns {number}
- */
-function systemStyleOrder (s1, s2) {
-	var s1KeyName = getCssKey(cssNames, s1.name);
-	var s2KeyName = getCssKey(cssNames, s2.name);
-
-	var s1Order = cssNames[s1KeyName] ? cssNames[s1KeyName].order : 9999;
-	var s2Order = cssNames[s2KeyName] ? cssNames[s2KeyName].order : 9999;
-	return s1Order - s2Order;
-}
+var styleset_utils = require('../lib/styleset-utils');
 
 function createStyleset(stylesheetName, jsonStyleset, next) {
 	var styleset = new Styleset({
@@ -149,7 +124,7 @@ function createStyleset(stylesheetName, jsonStyleset, next) {
 					//console.log('BEFORE');
 					//console.log(JSON.stringify(populatedStyleset.styles));
 
-					populatedStyleset.styles.sort(systemStyleOrder);
+					populatedStyleset.styles.sort(styleset_utils.systemStyleOrder);
 
 					//console.log('AFTER');
 					//console.log(JSON.stringify(populatedStyleset.styles));
